+++ CHANGELOG.md
@@ -68,6 +68,11 @@
 -   Added support for `cameraPosition` for the mapPortalBot and miniMapPortalBot.
 -   Added the `cameraMapPosition` tag for the mapPortalBot and miniMapPortalBot.
     -   Contains the longitude (x), latitude (y), and altitude (z) of the camera.
+-   Added the ability to automatically inject the bootstrap AUX from the server to reduce initialization time on the client.
+    -   Only supported on server deployments.
+    -   Set `serverInjectBootstrapper` to `true` on the web config in the server config.
+-   Added the ability to configure a custom favicon for server-based deployments.
+    -   Use the `icons` property on the web config in the server config.
 
 ### :bug: Bug Fixes